move all image files to its own folder
