+++ src/lib/background.png
