+++ src/style.css
@@ -4,7 +4,7 @@
 }
 
 body {
-  background-image: url('./background.png');
+  background-image: url('./lib/background.png');
   background-size: 250px;
   background-attachment: fixed;
 }
@@ -55,7 +55,7 @@ body {
 }
 
 .contacttitleimage {
-  background-image: url('./Contact_Banner.jpg');
+  background-image: url('./lib/Contact_Banner.jpg');
   background-repeat: no-repeat;
   background-size: cover;
   height: 400px;
